Secrets: Bow damage above 400k-ish should now work. Secrets: Negative AC values should no longer exist, as all AC values are now sint32s instead of the improper sint16. In some cases they were uint32, this was corrected as well. Secrets: Backstab damage above 400k-ish should now work. Secrets: Critical Damage above 20 million or so should work. Secrets: The above are hackish fixes and should probably have their internal datatypes be turned into int64's throughout the code eventually. This is what SOE does on the client, but not over the wire.
NOTE: If the Crit/Backstab/BowDmg/NegativeAC stuff cause issues, revert them!

+++ trunk/EQEmuServer/zone/zonedump.h
@@ -59,7 +59,7 @@ struct NPCType
 	int32	adventure_template;
 	int32	trap_template;
     int8    light;		//not loaded from DB
-	int16	AC;
+	int32	AC;
 	int32	Mana;	//not loaded from DB
 	int16	ATK;	//not loaded from DB
 	int16	STR;
